+++ core/src/org/rexellentgames/dungeon/entity/level/painter/TunnelPainter.java
@@ -22,32 +22,32 @@ public class TunnelPainter extends Painter {
 				if (door.x == room.left) {
 					from = room.left + 1;
 
-					for (int i = door.y; i != c.y; i += step) {
+					for (int i = (int) door.y; i != c.y; i += step) {
 						level.set(from, i, floor);
 					}
 				} else if (door.x == room.right) {
 					to = room.right - 1;
 
-					for (int i = door.y; i != c.y; i += step) {
+					for (int i = (int) door.y; i != c.y; i += step) {
 						level.set(to, i, floor);
 					}
 				} else {
 					if (door.x < from) {
-						from = door.x;
+						from = (int) door.x;
 					}
 
 					if (door.x > to) {
-						to = door.x;
+						to = (int) door.x;
 					}
 
-					for (int i = door.y + step; i != c.y; i += step) {
-						level.set(door.x, i, floor);
+					for (int i = (int) (door.y + step); i != c.y; i += step) {
+						level.set((int) door.x, i, floor);
 					}
 				}
 			}
 
 			for (int i = from; i <= to; i++) {
-				level.set(i, c.y, floor);
+				level.set(i, (int) c.y, floor);
 			}
 		} else {
 			int from = room.bottom - 1;
@@ -59,32 +59,32 @@ public class TunnelPainter extends Painter {
 				if (door.y == room.top) {
 					from = room.top + 1;
 
-					for (int i = door.x; i != c.x; i += step) {
+					for (int i = (int) door.x; i != c.x; i += step) {
 						level.set(i, from, floor);
 					}
 				} else if (door.y == room.bottom) {
 					to = room.bottom - 1;
 
-					for (int i = door.x; i != c.x; i += step) {
+					for (int i = (int) door.x; i != c.x; i += step) {
 						level.set(i, to, floor);
 					}
 				} else {
 					if (door.y < from) {
-						from = door.y;
+						from = (int) door.y;
 					}
 
 					if (door.y > to) {
-						to = door.y;
+						to = (int) door.y;
 					}
 
-					for (int i = door.x + step; i != c.x; i += step) {
-						level.set(i, door.y, floor);
+					for (int i = (int) (door.x + step); i != c.x; i += step) {
+						level.set(i, (int) door.y, floor);
 					}
 				}
 			}
 
 			for (int i = from; i <= to; i++) {
-				level.set(c.x, i, floor);
+				level.set((int) c.x, i, floor);
 			}
 		}
